增加从n张卡牌中选取m张的所有组合的算法
同时将程序的具体执行转移到run这个工具类中,main仅用来调用run中的方法
